[add] Select Input & Repository Select components [add] GitHub Organization model & OAuth middleware [optimize] split Article Editor component & page
